Refactor ContentInjector1 to use Dictionary.
Instead of using a list of asset names which is linear searched whenever
CanEdit and Edit are called, we can use a dictionary that maps asset
names to methods that inject assets. CanEdit need only check if the
asset name is present in the dictionary which is more efficient than
a linear search. Edit can then map the asset name to a funtion rather
than having so many if else's.

We used to use SMAPI's AssetNameEquals which would normalize the string
we sent it. Since we perform the comparisons, we need to normalize
our asset names. This is done when initializing the Dictionary using
SMAPI's NormalizeAssetName.