Redo how we fetch zoomLevel
According to Pathos, on Android `Game1.options` is a field. If we use `Game1.options` then the compiler -- based on SDV1.5 -- will use property-retrieval procedure and this will cause problems on Android.

So we use `Reflection.GetField` on Android, and use `Game1.game1.instanceOptions` on PC.

This means that to not do reflection multiple times, Adjust can no longer be an extension that lives outside the func; we "embed" it into `OnRenderingHud` so it can grab the already-retrieved `zoom` value directly from its containing func.