Complex options can now have dynamic height.
This change updates the Table object in SpaceShared with an optional
parameter to the constructor to specify whether the table uses fixed row
hight (the existing behavior and new default) or lays out elements based
on their Height (which may change).  It also tweaks the computation of
whether elements are on screen so that the last element on the screen will
be partially drawn rather than leaving some amount of blank space that
gets drawn in once the page scrolls.

The GMCM SpecificModConfigMenu enables the new Table feature, and the
Complex option widget now passes though the call to Height each time it is
asked rather than setting a fixed height at construction.