Update FixMagicIfNeeded
Function in message was missing a fix for skill points - connected to another bigger bug that I have no idea how to fix. 

In Multiplayer, players can be set to the level of the owner of the world due to unknown circumstances after unlocking magic. When this happens, though, the player whose level was set does not get the skill points and becomes unable to get those skill points that were skipped. Due to this, my assumption was that this was the function that was called, which is why I've added this change to at least give the players the correct amount of skill points.